requirements.txt
streamlit>=1.30.0
requests>=2.25.0
azure-identity>=1.15.0
openai>=1.0.0
azure-cognitiveservices-speech>=1.36.0
azure-ai-contentsafety>=1.0.0
langgraph>=0.1.0
langchain>=0.1.0
python-dotenv>=0.15.0
azure-keyvault-secrets>=4.7.0
